Add a bit of robustness to mail sending
We keep a pool of connections around. Some of them may be interrupted
for one reason or another and we have no real way to detect that other
than sending data over the wire. The simplest way to mitigate transient
failures like this is just to retry on a network failure.